+++ src/main/java/org/salemelrahal/jinn/test/api/NetworkTester.java
@@ -10,6 +10,6 @@ import org.salemelrahal.jinn.test.TrainingTest;
  * Used to determine the error of a network for a given TrainingTest or TrainingSuite.
  */
 public interface NetworkTester {
-	public BigDecimal test(Network network, TrainingSuite suite);
-	public BigDecimal test(Network network, TrainingTest test);
+	public double test(Network network, TrainingSuite suite);
+	public double test(Network network, TrainingTest test);
 }
